HMW-200 Removed the Temp Monitoring Clustering switch and updated the code so that clustering is automatically activated if there are more than 20 stations in the HUC.
