+++ app/client/src/components/shared/LocationMap.js
@@ -8,6 +8,7 @@ import React, {
   useState,
 } from 'react';
 import type { Node } from 'react';
+import { render } from 'react-dom';
 import { css } from 'styled-components/macro';
 import StickyBox from 'react-sticky-box';
 import { useNavigate } from 'react-router-dom';
@@ -34,7 +35,10 @@ import {
 } from 'utils/mapFunctions';
 import MapErrorBoundary from 'components/shared/ErrorBoundary.MapErrorBoundary';
 // contexts
-import { useFetchedDataDispatch, useFetchedDataState } from 'contexts/FetchedData';
+import {
+  useFetchedDataDispatch,
+  useFetchedDataState,
+} from 'contexts/FetchedData';
 import { LocationSearchContext } from 'contexts/locationSearch';
 import {
   useOrganizationsContext,
@@ -640,7 +644,7 @@ function LocationMap({ layout = 'narrow', windowHeight, children }: Props) {
         symbol: {
           type: 'simple-marker',
           style: 'circle',
-          color: colors.lightPurple(0.3),
+          color: colors.lightPurple(0.5),
         },
         visualVariables: [
           {
@@ -1125,7 +1129,10 @@ function LocationMap({ layout = 'narrow', windowHeight, children }: Props) {
   // updates the features on the monitoringStationsLayer
   useEffect(() => {
     if (!monitoringLocationsLayer) return;
-    if (!monitoringLocations.data.features || monitoringLocations.status !== 'success') {
+    if (
+      !monitoringLocations.data.features ||
+      monitoringLocations.status !== 'success'
+    ) {
       return;
     }
 
@@ -1182,6 +1189,58 @@ function LocationMap({ layout = 'narrow', windowHeight, children }: Props) {
         },
       });
     });
+
+    if (stationsSorted.length > 20) {
+      monitoringLocationsLayer.featureReduction = {
+        type: 'cluster',
+        clusterRadius: '100px',
+        clusterMinSize: '24px',
+        clusterMaxSize: '60px',
+        popupEnabled: true,
+        popupTemplate: {
+          title: 'Cluster summary',
+          content: (feature) => {
+            const content = (
+              <div style={{ margin: '0.625em' }}>
+                This cluster represents{' '}
+                {feature.graphic.attributes.cluster_count} stations
+              </div>
+            );
+
+            const contentContainer = document.createElement('div');
+            render(content, contentContainer);
+
+            // return an esri popup item
+            return contentContainer;
+          },
+          fieldInfos: [
+            {
+              fieldName: 'cluster_count',
+              format: {
+                places: 0,
+                digitSeparator: true,
+              },
+            },
+          ],
+        },
+        labelingInfo: [
+          {
+            deconflictionStrategy: 'none',
+            labelExpressionInfo: {
+              expression: "Text($feature.cluster_count, '#,###')",
+            },
+            symbol: {
+              type: 'text',
+              color: '#000000',
+              font: { size: 10, weight: 'bold' },
+            },
+            labelPlacement: 'center-center',
+          },
+        ],
+      };
+    } else {
+      monitoringLocationsLayer.featureReduction = undefined;
+    }
 
     monitoringLocationsLayer.queryFeatures().then((featureSet) => {
       monitoringLocationsLayer.applyEdits({
