+++ app/client/src/components/shared/LocationSearch.js
@@ -7,9 +7,7 @@ import React, {
   useRef,
   useState,
 } from 'react';
-import { render } from 'react-dom';
 import { css } from 'styled-components/macro';
-import Switch from 'components/shared/Switch';
 import { useNavigate } from 'react-router-dom';
 import FeatureLayer from '@arcgis/core/layers/FeatureLayer';
 import Locator from '@arcgis/core/tasks/Locator';
@@ -161,24 +159,6 @@ const searchBoxStyles = css`
   }
 `;
 
-/**
- * TODO
- *   Delete the following styles once the
- *   testing clustering switch is removed.
- */
-const tempClusteringContainerStyles = css`
-  display: flex;
-  justify-content: space-between;
-`;
-const tempClusteringLabelContainerStyles = css`
-  display: flex;
-  align-items: center;
-`;
-const tempClusteringLabelStyles = css`
-  margin: 0;
-  margin-right: 10px;
-`;
-
 type Props = {
   route: string,
   label: Node,
@@ -198,14 +178,7 @@ function LocationSearch({ route, label }: Props) {
   const sourceEnterPress = useKeyPress('Enter', sourceList);
   const clearButton = useRef();
   const clearEnterPress = useKeyPress('Enter', clearButton);
-  const {
-    searchText,
-    watershed,
-    huc12,
-    monitoringLocations,
-    monitoringLocationsLayer,
-    visibleLayers,
-  } = useContext(LocationSearchContext);
+  const { searchText, watershed, huc12 } = useContext(LocationSearchContext);
 
   const placeholder = 'Search by address, zip code, or place...';
   const [allSources] = useState([
@@ -738,9 +711,6 @@ function LocationSearch({ route, label }: Props) {
     };
   }, [suggestionsRef]);
 
-  // TODO Delete this state when the TEMP clustering switch is removed
-  const [clusteringEnabled, setClusteringEnabled] = useState(false);
-
   return (
     <>
       {errorMessage && (
@@ -749,87 +719,9 @@ function LocationSearch({ route, label }: Props) {
         </div>
       )}
 
-      {visibleLayers.monitoringLocationsLayer &&
-      monitoringLocationsLayer &&
-      monitoringLocations.status === 'success' &&
-      monitoringLocations.data.features.length ? (
-        <div css={tempClusteringContainerStyles}>
-          <label css={labelStyles} htmlFor="hmw-search-input">
-            {label}
-          </label>
-          <div css={tempClusteringLabelContainerStyles}>
-            <label css={tempClusteringLabelStyles}>
-              TEMP Monitoring Clustering
-            </label>
-            <Switch
-              checked={clusteringEnabled}
-              onChange={(checked) => {
-                setClusteringEnabled(checked);
-
-                if (checked) {
-                  monitoringLocationsLayer.renderer.symbol.color =
-                    colors.lightPurple(0.8);
-                  monitoringLocationsLayer.featureReduction = {
-                    type: 'cluster',
-                    clusterRadius: '100px',
-                    clusterMinSize: '24px',
-                    clusterMaxSize: '60px',
-                    popupEnabled: true,
-                    popupTemplate: {
-                      title: 'Cluster summary',
-                      content: (feature) => {
-                        const content = (
-                          <div style={{ margin: '0.625em' }}>
-                            This cluster represents{' '}
-                            {feature.graphic.attributes.cluster_count} stations
-                          </div>
-                        );
-
-                        const contentContainer = document.createElement('div');
-                        render(content, contentContainer);
-
-                        // return an esri popup item
-                        return contentContainer;
-                      },
-                      fieldInfos: [
-                        {
-                          fieldName: 'cluster_count',
-                          format: {
-                            places: 0,
-                            digitSeparator: true,
-                          },
-                        },
-                      ],
-                    },
-                    labelingInfo: [
-                      {
-                        deconflictionStrategy: 'none',
-                        labelExpressionInfo: {
-                          expression: "Text($feature.cluster_count, '#,###')",
-                        },
-                        symbol: {
-                          type: 'text',
-                          color: '#000000',
-                          font: { size: 10, weight: 'bold' },
-                        },
-                        labelPlacement: 'center-center',
-                      },
-                    ],
-                  };
-                } else {
-                  monitoringLocationsLayer.renderer.symbol.color =
-                    colors.lightPurple(0.3);
-                  monitoringLocationsLayer.featureReduction = undefined;
-                }
-              }}
-            />
-          </div>
-        </div>
-      ) : (
-        <label css={labelStyles} htmlFor="hmw-search-input">
-          {label}
-        </label>
-      )}
+      <label css={labelStyles} htmlFor="hmw-search-input">
+        {label}
+      </label>
 
       <form
         css={formStyles}
